+++ lib/settings/on_init_error_handler.dart
@@ -1,0 +1,1 @@
+typedef OnInitErrorHandler = void Function(String e);